backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
opencv-python
numpy
easyocr
